fixed #1573. I had test on my env. (#1629)
